Amélioration de la fonction fetchMoviesFromAPI : - Ajout de la vérification de la configuration de l'API (URL et clé manquantes). - Mise en place d'un système de cache avec revalidation toutes les heures pour optimiser les performances. - Validation des paramètres dynamiques avant ajout à l'URL (évite d'ajouter des valeurs nulles ou indéfinies). - Amélioration de la gestion des erreurs en fournissant des détails sur l'échec de la requête (status et message). - Meilleure gestion des paramètres par défaut dans l'URL.
Modification du nom du fichier et de l'architecture.
